# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cube-back",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/visibilites": {
            "get": {
                "tags": [
                    "Visibilite"
                ],
                "summary": "Récupérer toutes les visibilitées",
                "operationId": "bbd81e2d814427b8d99610f55bf5eb1b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page (1-100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for searching items",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort items by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Visibilite"
                ],
                "summary": "Créer une visibilité",
                "operationId": "0befd5008d8e5d9760ee73c7af4e208a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_vis"
                                ],
                                "properties": {
                                    "intitule_vis": {
                                        "type": "string",
                                        "maxLength": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/visibilites/{id}": {
            "get": {
                "tags": [
                    "Visibilite"
                ],
                "summary": "Récupérer une visibilité spécifique",
                "operationId": "f9bb654bc080c8728e715caeff7a39e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Visibilite"
                ],
                "summary": "Modifier une visibilité",
                "operationId": "602b12e30154361b248746d6861cdd35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_vis"
                                ],
                                "properties": {
                                    "intitule_vis": {
                                        "type": "string",
                                        "maxLength": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Visibilite"
                ],
                "summary": "Supprimer une visibilité",
                "operationId": "05d8b04f41bae782b98cc99e9582b882",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "AuthLogin"
                ],
                "summary": "Login",
                "operationId": "bc76a9d52929cab7a147ca7b5c527430",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "AuthLogin"
                ],
                "summary": "Logout",
                "operationId": "0ab4e3cf7bc4f7ec81b48f7928c7de7c",
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}